+++ src/base/image_loader.c
@@ -21,10 +21,9 @@
 
 #include "base/image_loader.h"
 
-ret_t image_loader_load(image_loader_t* loader, const uint8_t* buff, uint32_t size,
-                        bitmap_t* bitmap) {
-  return_value_if_fail(loader != NULL && loader->load != NULL && buff != NULL && bitmap != NULL,
+ret_t image_loader_load(image_loader_t* loader, const asset_info_t* asset, bitmap_t* bitmap) {
+  return_value_if_fail(loader != NULL && loader->load != NULL && asset != NULL && bitmap != NULL,
                        RET_BAD_PARAMS);
 
-  return loader->load(loader, buff, size, bitmap);
+  return loader->load(loader, asset, bitmap);
 }