+++ src/base/image_loader.h
@@ -23,13 +23,14 @@
 #define TK_IMAGE_LOADER_H
 
 #include "base/bitmap.h"
+#include "base/assets_manager.h"
 
 BEGIN_C_DECLS
 
 struct _image_loader_t;
 typedef struct _image_loader_t image_loader_t;
 
-typedef ret_t (*image_loader_load_t)(image_loader_t* loader, const uint8_t* buff, uint32_t size,
+typedef ret_t (*image_loader_load_t)(image_loader_t* loader, const asset_info_t* asset,
                                      bitmap_t* bitmap);
 
 /**
@@ -44,13 +45,11 @@ struct _image_loader_t {
  * @method image_loader_load
  * 加载图片。
  * @param {image_loader_t*} loader loader对象。
- * @param {uint8_t*} buff 图片数据。
- * @param {uint32_t} size 图片数据长度。
+ * @param {const asset_info_t*} asset 资源。
  * @param {bitmap_t*} bitmap 用于返回位图对象。
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
-ret_t image_loader_load(image_loader_t* loader, const uint8_t* buff, uint32_t size,
-                        bitmap_t* bitmap);
+ret_t image_loader_load(image_loader_t* loader, const asset_info_t* asset, bitmap_t* bitmap);
 
 END_C_DECLS