+++ src/ext_widgets/ext_widgets.c
@@ -23,6 +23,7 @@
 #include "switch/switch.h"
 #include "guage/guage.h"
 #include "guage/guage_pointer.h"
+#include "gif_image/gif_image.h"
 #include "svg_image/svg_image.h"
 #include "keyboard/keyboard.h"
 #include "base/widget_factory.h"
@@ -67,6 +68,7 @@ ret_t tk_ext_widgets_init() {
   widget_factory_register(widget_factory(), WIDGET_TYPE_IMAGE_ANIMATION, image_animation_create);
   widget_factory_register(widget_factory(), WIDGET_TYPE_PROGRESS_CIRCLE, progress_circle_create);
   widget_factory_register(widget_factory(), WIDGET_TYPE_SVG_IMAGE, svg_image_create);
+  widget_factory_register(widget_factory(), WIDGET_TYPE_GIF_IMAGE, gif_image_create);
 
   return RET_OK;
 }