+++ src/ext_widgets/gif_image/gif_image.h
@@ -1,7 +1,7 @@
 ﻿/**
- * File:   svg_image.h
+ * File:   gif_image.h
  * Author: AWTK Develop Team
- * Brief:  svg_image
+ * Brief:  gif_image
  *
  * Copyright (c) 2018 - 2018  Guangzhou ZHIYUAN Electronics Co.,Ltd.
  *
@@ -15,12 +15,12 @@
 /**
  * History:
  * ================================================================
- * 2018-11-22 Li XianJing <xianjimli@hotmail.com> created
+ * 2018-11-25 Li XianJing <xianjimli@hotmail.com> created
  *
  */
 
-#ifndef TK_SVG_IMAGE_H
-#define TK_SVG_IMAGE_H
+#ifndef TK_GIF_IMAGE_H
+#define TK_GIF_IMAGE_H
 
 #include "base/widget.h"
 #include "base/image_base.h"
@@ -28,21 +28,24 @@
 BEGIN_C_DECLS
 
 /**
- * @class svg_image_t
+ * @class gif_image_t
  * @parent widget_t
  * @annotation ["scriptable"]
- * SVG图片控件。
+ * GIF图片控件。
  */
-typedef struct _svg_image_t {
+typedef struct _gif_image_t {
   image_base_t image_base;
 
   /*private*/
-  const asset_info_t* bsvg_asset;
-} svg_image_t;
+  int* delays;
+  uint32_t index;
+  uint32_t timer_id;
+  uint32_t frames_nr;
+} gif_image_t;
 
 /**
- * @method svg_image_create
- * 创建svg_image对象
+ * @method gif_image_create
+ * 创建gif_image对象
  * @annotation ["constructor", "scriptable"]
  * @param {widget_t*} parent 父控件
  * @param {xy_t} x x坐标
@@ -52,22 +55,22 @@ typedef struct _svg_image_t {
  *
  * @return {widget_t*} 对象。
  */
-widget_t* svg_image_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h);
+widget_t* gif_image_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h);
 
 /**
- * @method svg_image_cast
- * 转换为svg_image对象(供脚本语言使用)。
+ * @method gif_image_cast
+ * 转换为gif_image对象(供脚本语言使用)。
  * @annotation ["cast", "scriptable"]
- * @param {widget_t*} widget svg_image对象。
+ * @param {widget_t*} widget gif_image对象。
  *
- * @return {widget_t*} svg_image对象。
+ * @return {widget_t*} gif_image对象。
  */
-widget_t* svg_image_cast(widget_t* widget);
+widget_t* gif_image_cast(widget_t* widget);
 
-#define WIDGET_TYPE_SVG_IMAGE "svg"
+#define WIDGET_TYPE_GIF_IMAGE "gif"
 
-#define SVG_IMAGE(widget) ((svg_image_t*)(widget))
+#define GIF_IMAGE(widget) ((gif_image_t*)(widget))
 
 END_C_DECLS
 
-#endif /*TK_SVG_IMAGE_H*/
+#endif /*TK_GIF_IMAGE_H*/