+++ src/ext_widgets/svg_image/svg_image.c
@@ -15,7 +15,7 @@
 /**
  * History:
  * ================================================================
- * 2018-01-28 Li XianJing <xianjimli@hotmail.com> created
+ * 2018-11-22 Li XianJing <xianjimli@hotmail.com> created
  *
  */